Introduce basics for Parsers and Printers.
This change introduces basic foundation for jribble parsers and printers.
Parser is (almost) already capable of parsing jrrible output for Hello
sample which is covered by JribbleSamplesTestCase. The only modification
that has been applied to that jribble source was to replace "super()" call
with "this.super()" call. This change is planned for jribble's syntax anyway.

The approach for testing parsers is to test for existence of correct left
inverse. Parser 'par' can be seen as function (String => T) which has
corresponding Printer 'pr' which can be seen as function (T => String).
Now, all we need to check is if
  (par compose pr) == identity[T].

By definition of function equality we need to check the equation above for
every element of the domain T. In principle this is not decidable, so what
we do is sampling. We introduce Arbitrary[T] which is generator of random
samples of T and we use it for testing that equality.

Obviously, if both Parser and Printer have a bug of the same kind so their
composition yields identity we won't detect it. However, in practice it's
ver unlikely to have such combination.

Additionally, it's worth noting why we don't test for right inverse of
Printer. We face two problems. First of all, Printer is not determined
uniquely. In plain words, you can make many formatting choices when trying
to define Printer and in all such cases you'll get correct Printer.
Secondly, there is no way to define Arbitrary[String] so testing for
function equality will yield meaningful results.

The above describes the approach for testing both parsers and
printers that can be found in PrintersParsersTestCase. ScalaCheck is used
as a library making it easy to define Arbitrary instances through monadic
composition of generators.

This approach gives a very good and systematic coverage of our parsers
and printers.

Also added JribbleSampleTestCase mentioned earlier which is just JUnit
test. Added dependency on junit and it's interface for sbt to project
description.

Signed-off-by: Grzegorz Kossakowski <grek@google.com>
Change-Id: I36a5602c546d13f2a9bd5239f53e14ff9e474f09