Use JSNI syntax for types and method signatures.
Instead of using Java-like syntax for primitive types and reference types
use one established by JSNI. This is done uniformly, so variables and
parameters use the same syntax.

Signature contains method's full coordinates which includes class name
where it was defined and it's name. This means that jribble syntax
contains engough information to make method lookups static and not bother
about virtual calls. Actually, this is the main purpose of this change and
JSNI was chosen because it encodes all needed information and is used in
GWT.

As consequence there were a few refactorings applied:
  * ClassName was renamed to Ref.
  * MethodCall and StaticMethodCall don't have name property anymore
    (it's covered by Signature)
  * NewCall doesn't have classRef property anymore (again covered by
    Signature)

Change-Id: I98e50c0d22150c00fe1a03c259eab762f7a4decf
Signed-off-by: Grzegorz Kossakowski <grek@google.com>