Adding jars from lib_managed directories to the repository.
We are relying on snapshot versions of libraries like scalacheck so in
order to guarantee reproducible builds we should have all jars checked-in.

The version of scalacheck library we include in this commit includes bug
fixes that made it apparent that jribble parser is buggy. The bug in
scalacheck that was hiding bugs in our parser is described here:
http://code.google.com/p/scalacheck/issues/detail?id=72

In short this bug caused scalacheck generators to generate only a small
subset of possible input data thus not really covering all possible cases.

We'll address these problems in subsequent commits.

Change-Id: I3454b285a0cc16384885a01369adda8cf5769d1c
Signed-off-by: Grzegorz Kossakowski <grek@google.com>