Removed some boilerplate from Parsers thanks to implicit conversions.
It was a very repeating pattern to convert nested datastructure returned
by ~ parser combinator into case class. If the order and number of
arguments to case class is the same as number of arguments returned by
~ combinators then we can implicitly perform the conversion.

Conversions we are using are defined in ImplicitConversions trait that
comes with Scala's parser combinator library.

Change-Id: I2ca2e1933c5da2e92087dbc638217f8ba4d6e20e
Signed-off-by: Grzegorz Kossakowski <grek@google.com>