Implemented a few basic Shrinkers for Jribble AST nodes.
ScalaCheck includes concept of Shrink which is a function T => Stream[T]
with property that Stream has this property that next element in the
Stream is not bigger (or not more complex) than previous one.

Shrink is being used once ScalaCheck find a sample that falsifies the
tested property and uses it to come up with smaller sample still
falsifying the property.

Not all AST nodes have their corresponding Shrinks. More of them will be
added later on.

Change-Id: I85f07286d22772b315970d729af0dddbf3699d05
Signed-off-by: Grzegorz Kossakowski <grek@google.com>