Added newer version of scala-library-jribble.jar
Added latest version of a subset of Scala Library compiled
to jribble. This subset consisting of 1785 jribble files is
being parsed successfully with current jribble parser.

This version of scala-library-jribble.jar file cannot be
compiled with any committed version of a scalac
because scalac has been hacked in ugly ways in order
to produce this jar. Once better solution for scalac is
found we'll have commit allowing to produce that jar.

Change-Id: I82eca881f4a252a6d15b08404e8d5873d848c599
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>